[Leveling] el jugador puede subir de nivel

+++ Assets/Scripts/Enemies/Enemy.cs
@@ -205,7 +205,7 @@ namespace Enemies
         private void OnDestroy()
         {
             OnEnemyDeath?.Invoke(gameObject);
-            PlayerStats._instance.gold += gold;
+            PlayerStats._instance.gold += (int)gold;
             WaveController._instance.EnemyDeath();
         }
